HID-C: Add missing implementation for get report handling
This patch adds the missing implementation for responding to GET Report
requests from HID Host for different report types. For Input report,
handling is added for sending default values if report id is not present,
for output reports, reports are now stored and error is reported if report
id is invalid otherwise default values are returned.

Change-Id: I3bdc8538bfce760959166f9ecaf6582b750f5381
CRs-Fixed: 593100