BT: Fix to close FTP service while BT turning OFF in race condition
A case, where automation is running for many FTP use cases, Some of
the test cases are doing  BT OFF/ON while running the test case.
Sometimes FTP service is not closed, when STATE_TURNING_OFF indent
is received by FTP service. But adapter state was already moved to
STATE_OFF. As FTP service is not closed, So SDP registration for
FTP does not happen on next BT TURN ON. This patch will fix the
issue.

Change-Id: I903385224f9a76dccff524c3e1ed1c0a912b6137
CRs-Fixed: 653213